+++ ewm-main-service/src/main/java/ru/practicum/category/api/dto/NewCategoryDto.java
@@ -1,0 +1,22 @@
+package ru.practicum.category.api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class NewCategoryDto {
+    private @NotNull @Size(
+            max = 50,
+            min = 1
+    ) String name;
+}
