+++ ewm-main-service/src/main/java/ru/practicum/compilation/api/dto/NewCompilationDto.java
@@ -1,15 +1,21 @@
-package ru.practicum.compilation.dto;
+package ru.practicum.compilation.api.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.List;
 import javax.validation.constraints.NotNull;
 
+@Builder
+@Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class NewCompilationDto {
-    private List<Long> events;
-    private boolean pinned;
     private @NotNull String title;
+    private Boolean pinned;
+    private List<Long> events;
 }
